requirements.txt
Django==3.1.*
environs[django]==9.3.0
Pillow==8.0.*  # required by Windows environment
django-extensions==3.2.3
django-debug-toolbar==4.2.0
Pygments==2.16.1